feat: add gnome dconf module and switch zippyrus to gnome
replace kde plasma with gnome because the new kde plasma is honestly pretty
bad. gnome has better defaults and is configurable with dconf, and the
gsettings configuration is a first-class citizen on nix, unlike any of the
plasma configuration. switching to gnome makes our setup a bit more
reproducible using nix.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>